==> 1_PROCESSING/0_clean_cities_data.ipynb <==
import pandas as pd

df = pd.read_json("../files/raw/cities_col.json")
print(len(df[df["location"].isna()]))

ciudades= df[df["location"].isna()][["name","department_name"]]
manual_map={'Tarapacá':(2.557320, -69.78275),
 'Puerto Arica':(-2.148599, -71.75411),
 'Puerto Alegría':(-1.005497, -74.01443),
 'Miriti- Paraná':(-0.7239085, -71.10315),
 'La Victoria':(-0.09504421, -71.15686),
 'La Pedrera':(-1.320789, -69.57810),
 'La Chorrera':(-1.442697, -72.78927),
 'El Encanto':(-1.747107, -73.20959),
 'Ciudad Bolívar':(5.849018, -76.02022),
 'Belén de Bajirá':(7.371074, -76.71429),
 'Patía':(2.069477, -77.05323),
 'Tuchín':[8.551022, -75.85167],
 'Barranco Minas':(3.494698, -69.81403),
 'Mapiripana':(2.996659, -70.21850),
 'Cácota':(7.268399, -72.64222),
 'Herrán':(7.506128, -72.48341),
 'Armero-Guayabal':(5.031632, -74.88488),
 'Calima':(3.925724, -76.49444),
 'Caruru':(1.012148, -71.29496),
 'Apulo':(4.519548, -74.59564),
 'Ospina Perez':(4.088817, -74.47739),
 'Rafael Reyes':(4.520464, -74.59511),
 'El Bordo':(2.115983, -76.98311),
 'Puerto Bolivar':(12.25474, -71.96539),
 'Acaricuara':(0.6979801, -70.23276),
 'Villa Fatima':(0.9731719, -69.94996),
 'Coloto':(3.028130, -76.40616),
 'Iscuande':(2.450622, -77.97927),
 'Popayán':(2.439376, -76.61367),
 'El Darién':(3.922297, -76.49306),
 'Puerto Santander':(-0.6212877, -72.38332),
 'Venecia':(4.089466, -74.47779),
 'Otra (Vichada)':(5.836529669,-68.141222336),
 'Otra (Vaupes)':(1.06148231,-70.46688401),
 'Otra (Tolima)':(4.451921427,-75.252590995),
 'Otra (Sucre)':(9.316673928,-75.431749665),
 'Otra (Santander)':(7.155833544,-73.111569971),
 'Otra (Risaralda)':(4.803662773,-75.795790934),
 'Otra (Quindio)':(4.499501128,-75.724899852),
 'Otra (Putumayo)':(1.26070915,-76.665843491)	,
 'Otra (Narino)':(1.083605497,-77.206100743),
 'Otra (Meta)':(4.09166877,-73.492915945),
 'Otra (Magdalena)':(11.121894357,-73.885278551),
 'Otra (La Guajira)':(11.242972141,-72.958763152),
 'Otra (Huila)':(2.993360046,-75.272362322),
 'Otra (Guainia)':(3.516363079,-68.008347279),
 'Otra (Cundinamarca)':(4.316107698,-74.181072702),
 'Otra (Cordoba)':(8.584698016,-75.950550788),
 'Otra (Choco)':(5.7543229,-76.65015147),
 'Otra (Cesar)':(10.218902768,-73.457847638),
 'Otra (Casanare)':(5.242744761,-72.25802857),
 'Otra (Caqueta)':(1.749138704,-75.558239064),
 'Otra (Caldas)':(5.08343301,-75.50727881),
 'Otra (Boyaca)':(5.518472615,-73.378018556),
 'Otra (Bolivar)':(10.463433617,-75.45889915),
 'Otra (Atlantico)':(10.98152051,-74.827715439),
 'Otra (Arauca)':(6.796280825,-70.509211529),
 'Otra (Antioquia)':(6.257590259,-75.611031065),
 'Otra (Amazonas)':	(-3.530058784,-70.04513691),
 'Otra (Valle del Cauca)':(3.399043723,-76.576492589),
 'Otra (Norte de Santander)':(7.270560088,-72.652121526),
 'Otra (Guaviare)':(2.484286091,-71.919169926),
 'Otra (Cauca)':(2.439376, -76.61367),
 'Otra (Bogotá D.C.)':(4.316107698,-74.181072702)}


df["location"]=df.apply(lambda city: str(manual_map[city["name"]]) if city["name"] in manual_map else city["location"], axis=1)

print(len(df[df["location"].isna()]))
df = df[~df["location"].isna()]

df.to_json("../files/raw/cities_col.json")
