==> 1_PROCESSING/2_applications_to_raw.ipynb <==
import pandas as pd
import glob

files = glob.glob("../files/batches_applications/data_download/*")
files

frames = [pd.read_parquet(file) for file in files]
applications = pd.concat(frames)

applications=applications.rename(columns={0:"application_id",1:"candidate_id",
                             2:"vacant_id",3:"application_status",4:"created_at",
                             5:"stage_id",6:"stage_order",7:"stage_name",8:"stage_entered_at"})


applications["application_status"].unique()

applications["application_status"]=applications["application_status"].map({"descartada":1,  'abandonó':0, 'activa':0, 'eliminada':0})

applications= applications[['application_id', 'candidate_id', 'vacant_id',  'application_status',
       'stage_id', 'stage_name', 'stage_entered_at', "stage_order", "created_at"]]
len(applications)

wit_apps = applications["candidate_id"].unique()
wit_apps = pd.DataFrame(wit_apps, columns=["id"])
wit_apps.to_json("../files/orginal_files/candidates/candidates_with_apps.json")

applications.reset_index(drop=True).to_json("../files/raw/applications.json")
