==> 1_PROCESSING/4_select_sample_size.ipynb <==
import pandas as pd 

candidates = pd.read_json("../files/raw/candidates.json")
candidates = candidates.sample(frac=0.3, random_state=123)

candidates.to_json("../files/sample_for_capacity/candidates.json")


candidates=candidates["candidate_id"].unique()

applications = pd.read_json("../files/raw/applications.json")

applications["candidate_id"][0]

candidates[1]

applications=applications[applications["candidate_id"].isin(candidates)]

applications.to_json("../files/sample_for_capacity/applications.json")
